GROUP 2 PAPER REVIEW OCT/requirements.txt
av==9.2.0
click==8.1.3
dlib==19.24.0
face_recognition_models==0.3.0
joblib==1.1.0
numpy==1.23.4
pandas<1.2.0
opencv_python==4.6.0.66
Pillow==9.2.0
scikit_learn==1.1.2
streamlit_webrtc==0.44.0
click ==8.1.3
pillow ==9.2.0
